requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
tqdm==4.65.0 